Enhancements and bug fixes - Use `refechQueries` to fixing client-side caching issue after mutations. - Use state for input values in `TaskCard` - Return full `task` objects from server-side mutations. - Standardize sytax across queries in QueryService. - Enhance README per requirements.

+++ app/graphql/mutations/destroy_task.rb
@@ -8,6 +8,9 @@ module Mutations
 
     def resolve(id:)
       task = Task.find(id)
+
+      return unless task
+
       task.destroy
         {
           id: id,
